Fix parameter passing in bin/coreir
Fixes https://github.com/leonardt/pycoreir/issues/100

The issue was that the parameters passed to the python script were not
being properly escaped when passed through to the shell command that
invokes the coreir binary. This uses the subprocess interface which
handles the escaping implicitly (and is a much safer/robust
implementation).