Explicitly load coreir libraries relative to binary path
Suppose you installed coreir to somewhere non-standard (/opt/coreir,
for example) and added the relevant bin directory (/opt/coreir/bin) to
your PATH.

Then the existing code would set COREIR_BINARY_PATH to
/opt/coreir/bin/coreir (good!) and then just try to load
'libcoreir.so' or similar. It should load
/opt/coreir/lib/libcoreir.so.

This isn't quite as useful as you might hope, because of coreir issue
number #851[1], but it gets most of the way (as a work-around for the
coreir bug, it suffices to leave the coreir build directory
uncleaned!).

[1] https://github.com/rdaly525/coreir/issues/851